requirements.txt
# Core FastAPI & server
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1

# Document processing
PyMuPDF==1.26.3
python-docx==1.2.0
docx2txt==0.9

# PDF and file utils
requests==2.32.4
python-multipart==0.0.20

# Embeddings & retrieval
faiss-cpu==1.12.0
sentence-transformers==5.0.0

# LangChain ecosystem
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-groq==0.3.7
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.9

# LLM APIs
openai==1.98.0
groq==0.31.0
google-generativeai==0.8.5

# Data validation & settings
pydantic==2.11.7
pydantic-settings==2.10.1

# Utilities
tiktoken==0.9.0
python-dotenv==1.1.1
jinja2==3.1.4